+++ test/HatTrick.DbEx.MsSql.Test.Database/Executor/_Functions/Minimum.cs
@@ -13,7 +13,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
     {
         [Theory]
         [MsSqlVersions.AllVersions]
-        public void Does_selecting_minimum_total_purchase_amount_succeed(int version, decimal expected = 5.11m)
+        public void Does_selecting_minimum_total_purchase_amount_succeed(int version, double expected = 5.11)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -23,7 +23,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 ).From(dbo.Purchase);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -31,7 +31,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
 
         [Theory]
         [MsSqlVersions.AllVersions]
-        public void Does_selecting_minimum_distinct_total_purchase_amount_succeed(int version, decimal expected = 5.11m)
+        public void Does_selecting_minimum_distinct_total_purchase_amount_succeed(int version, double expected = 5.11)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -41,7 +41,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 ).From(dbo.Purchase);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -50,7 +50,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "ORDER BY")]
-        public void Can_order_by_minimum_of_total_purchase_amount_ascending_succeed(int version, decimal expected = 5.11m)
+        public void Can_order_by_minimum_of_total_purchase_amount_ascending_succeed(int version, double expected = 5.11)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -61,7 +61,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 .OrderBy(db.fx.Min(dbo.Purchase.TotalPurchaseAmount));
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -70,7 +70,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "ORDER BY")]
-        public void Can_order_by_minimum_of_total_purchase_amount_descending_succeed(int version, decimal expected = 5.11m)
+        public void Can_order_by_minimum_of_total_purchase_amount_descending_succeed(int version, double expected = 5.11)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -81,7 +81,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 .OrderBy(db.fx.Min(dbo.Purchase.TotalPurchaseAmount).Desc);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -90,7 +90,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "ORDER BY")]
-        public void Can_order_by_minimum_of_total_purchase_amount_ascending_and_aliasing_succeed(int version, decimal expected = 5.11m)
+        public void Can_order_by_minimum_of_total_purchase_amount_ascending_and_aliasing_succeed(int version, double expected = 5.11)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -101,7 +101,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 .OrderBy(db.fx.Min(dbo.Purchase.TotalPurchaseAmount));
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -110,7 +110,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "ORDER BY")]
-        public void Can_order_by_minimum_of_total_purchase_amount_descending_and_aliasing_succeed(int version, decimal expected = 5.11m)
+        public void Can_order_by_minimum_of_total_purchase_amount_descending_and_aliasing_succeed(int version, double expected = 5.11)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -121,7 +121,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 .OrderBy(db.fx.Min(dbo.Purchase.TotalPurchaseAmount).Desc);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
