+++ test/HatTrick.DbEx.MsSql.Test.Database/Executor/_Functions/Sum.cs
@@ -13,7 +13,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
     {
         [Theory]
         [MsSqlVersions.AllVersions]
-        public void Does_sum_of_total_purchase_amount_succeed(int version, decimal expected = 308.15m)
+        public void Does_sum_of_total_purchase_amount_succeed(int version, double expected = 308.15)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -23,7 +23,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 ).From(dbo.Purchase);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -31,7 +31,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
 
         [Theory]
         [MsSqlVersions.AllVersions]
-        public void Does_sum_of_distinct_total_purchase_amount_succeed(int version, decimal expected = 299.15m)
+        public void Does_sum_of_distinct_total_purchase_amount_succeed(int version, double expected = 299.15)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -41,7 +41,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 ).From(dbo.Purchase);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -51,7 +51,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "ORDER BY")]
-        public void Can_order_by_sum_of_total_purchase_amount_ascending_succeed(int version, decimal expected = 308.15m)
+        public void Can_order_by_sum_of_total_purchase_amount_ascending_succeed(int version, double expected = 308.15)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -62,7 +62,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 .OrderBy(db.fx.Sum(dbo.Purchase.TotalPurchaseAmount));
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -71,7 +71,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "ORDER BY")]
-        public void Can_order_by_sum_of_total_purchase_amount_descending_succeed(int version, decimal expected = 308.15m)
+        public void Can_order_by_sum_of_total_purchase_amount_descending_succeed(int version, double expected = 308.15)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -82,7 +82,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 .OrderBy(db.fx.Sum(dbo.Purchase.TotalPurchaseAmount).Desc);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -91,7 +91,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "ORDER BY")]
-        public void Can_order_by_sum_of_total_purchase_amount_ascending_and_aliasing_succeed(int version, decimal expected = 308.15m)
+        public void Can_order_by_sum_of_total_purchase_amount_ascending_and_aliasing_succeed(int version, double expected = 308.15)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -102,7 +102,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 .OrderBy(db.fx.Sum(dbo.Purchase.TotalPurchaseAmount));
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -111,7 +111,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
         [Theory]
         [MsSqlVersions.AllVersions]
         [Trait("Operation", "ORDER BY")]
-        public void Can_order_by_sum_of_total_purchase_amount_descending_and_aliasing_succeed(int version, decimal expected = 308.15m)
+        public void Can_order_by_sum_of_total_purchase_amount_descending_and_aliasing_succeed(int version, double expected = 308.15)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -122,7 +122,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
                 .OrderBy(db.fx.Sum(dbo.Purchase.TotalPurchaseAmount).Desc);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
